Style/style code (#134)
* style: :art: Style solidity code

Style solidity code

* style: Add solhint and style code

Add solhint and styled the code to have a clean solhint check. Disable in the solhint.json some checks that we don't require and some solhint disable lines added to the solidity code.

* style: Add comments for solhint and improved naming

Add comments for solhint-disable. Also improve names for TokenVesting and StakeManager require strings.

* ci: Add ci solhint check

Add ci solhint check

* ci: Add install command to ci solhint

Add install command to ci solhint

* ci: Run solhint with npx

Run solhint with npx

* ci: Force a failure on CI if solhint raises any warnings

Force a failure on CI if solhint raises any warnings

+++ contracts/FLIP.sol
@@ -1,33 +1,35 @@
 pragma solidity ^0.8.0;
 
-
 import "@openzeppelin/contracts/token/ERC777/ERC777.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
 import "./interfaces/IFLIP.sol";
 import "./abstract/Shared.sol";
 
-
 /**
-* @title    FLIP contract
-* @notice   The FLIP utility token which is used to stake in the FLIP system and pay for
-*           trap fees with
-* @author   Quantaf1re (James Key)
-*/
+ * @title    FLIP contract
+ * @notice   The FLIP utility token which is used to stake in the FLIP system and pay for
+ *           trap fees with
+ * @author   Quantaf1re (James Key)
+ */
 contract FLIP is ERC777, Ownable, Shared {
-
     constructor(
         string memory name,
         string memory symbol,
         address[] memory defaultOperators,
         address receiver,
         uint256 mintAmount
-    ) ERC777(name, symbol, defaultOperators) Ownable() nzAddr(receiver) nzUint(mintAmount) {
+    )
+        ERC777(name, symbol, defaultOperators)
+        Ownable()
+        nzAddr(receiver)
+        nzUint(mintAmount)
+    {
         _mint(receiver, mintAmount, "", "");
     }
 
     function mint(
         address receiver,
-        uint amount,
+        uint256 amount,
         bytes memory userData,
         bytes memory operatorData
     ) external nzAddr(receiver) nzUint(amount) onlyOwner {
